+++ beacon-chain/core/feed/state/BUILD.bazel
