+++ beacon-chain/core/feed/state/notifier.go
@@ -1,4 +1,4 @@
-package statefeed
+package state
 
 import "github.com/prysmaticlabs/prysm/shared/event"
